(refactor): Analysis framework moved to its own mod
