+++ crates/cairo-lang-lowering/src/lib.rs
@@ -2,6 +2,7 @@
 //!
 //! This crate is responsible for handling the lowering phase.
 pub mod add_withdraw_gas;
+pub mod analysis;
 pub mod borrow_check;
 pub mod cache;
 pub mod concretize;
