+++ crates/cairo-lang-sierra-generator/src/local_variables.rs
@@ -17,8 +17,8 @@ use cairo_lang_utils::ordered_hash_set::OrderedHashSet;
 use cairo_lang_utils::unordered_hash_map::UnorderedHashMap;
 use cairo_lang_utils::unordered_hash_set::UnorderedHashSet;
 use itertools::{chain, zip_eq};
+use lowering::analysis::{Analyzer, BackAnalysis, StatementLocation};
 use lowering::borrow_check::Demand;
-use lowering::borrow_check::analysis::{Analyzer, BackAnalysis, StatementLocation};
 use lowering::borrow_check::demand::DemandReporter;
 use lowering::{Lowered, MatchInfo, Statement, VarRemapping, VarUsage};
 use salsa::Database;
